Rename `lib/sdk` to `lib/client-sdk` in preparation for the server SDK
Change-type: major
