+++ apps/ui/core/sdk.js
@@ -6,7 +6,7 @@
 
 import {
 	getSdk
-} from '../../../lib/sdk'
+} from '../../../lib/client-sdk'
 import * as environment from '../environment'
 
 export const sdk = getSdk({
