+++ lib/client-sdk/auth.js
