+++ lib/client-sdk/index.spec.js
