+++ lib/client-sdk/stream.js
