--- lib/sdk/DESCRIPTION.markdown
@@ -1,3 +1,0 @@
-The sdk is a client side library to interact with the Jellyfish infrastructure
-through its public interfaces (i.e. HTTP). Its meant to provide high level
-useful functionality to the web UI and any other clients.
