+++ package.json
@@ -32,7 +32,7 @@
       "lib/assert",
       "lib/ui-components",
       "lib/uuid",
-      "lib/sdk",
+      "lib/client-sdk",
       "lib/sync",
       "lib/worker",
       "lib/mail"
