+++ test/e2e/server/actions/action-increment.spec.js
@@ -5,7 +5,7 @@
  */
 
 const ava = require('ava')
-const helpers = require('../../sdk/helpers')
+const helpers = require('../../client-sdk/helpers')
 
 ava.serial.before(helpers.before)
 ava.serial.after(helpers.after)
