+++ test/e2e/server/actions/action-maintain-contact.spec.js
@@ -6,7 +6,7 @@
 
 const ava = require('ava')
 const uuid = require('uuid/v4')
-const helpers = require('../../sdk/helpers')
+const helpers = require('../../client-sdk/helpers')
 
 ava.serial.before(helpers.before)
 ava.serial.after(helpers.after)
