+++ test/e2e/server/index.spec.js
@@ -8,7 +8,7 @@ const ava = require('ava')
 const Bluebird = require('bluebird')
 const uuid = require('uuid/v4')
 const _ = require('lodash')
-const helpers = require('../sdk/helpers')
+const helpers = require('../client-sdk/helpers')
 const environment = require('../../../lib/environment')
 
 ava.serial.before(helpers.before)
