+++ test/e2e/server/messages.spec.js
@@ -5,7 +5,7 @@
  */
 
 const ava = require('ava')
-const helpers = require('../sdk/helpers')
+const helpers = require('../client-sdk/helpers')
 
 ava.serial.before(helpers.before)
 ava.serial.after(helpers.after)
