+++ test/e2e/server/users.spec.js
@@ -8,7 +8,7 @@ const ava = require('ava')
 const md5 = require('blueimp-md5')
 const nock = require('nock')
 const uuid = require('uuid/v4')
-const helpers = require('../sdk/helpers')
+const helpers = require('../client-sdk/helpers')
 
 ava.serial.before(helpers.before)
 ava.serial.after(helpers.after)
