+++ test/e2e/sync/helpers.js
@@ -5,7 +5,7 @@
  */
 
 const uuid = require('uuid/v4')
-const helpers = require('../sdk/helpers')
+const helpers = require('../client-sdk/helpers')
 
 exports.mirror = {
 	before: async (test) => {
